Add The Ability To Specify A Different MIME Type (#14)
There are many different mime types for multipart
requests and this change provides the option to
specify a different one.

If a multipart mime type is not specified the
middleware will default to multipart/form-data.

This change also changes the structure of the
middleware a bit. It changes the parent from
UrlEncoded to inherit from Middleware directly
instead. This is because we were overriding most
of the UrlEncoded methods anyway so bringing last
remaining methods over and removing that link in
the dependency chain seemed to make sense.

Another structural code change was to move the
methods that didn't need to be public under the
private keyword. The intent of this was to help
future developers know what they can change
without worrying breaking any external
dependencies.